+++ cabal.project
@@ -1,5 +1,5 @@
 packages:
+  ./timeline-core
   ./timeline
   ./timeline-hedgehog
-  ./timeline-tests
 tests: True
