+++ timeline-core/CHANGELOG.md
@@ -1,0 +1,4 @@
+# Changelog
+
+## 0.1.0.0
+- Open source the timeline library we use internally at Bellroy
