+++ timeline-core/timeline-core.cabal
@@ -1,0 +1,35 @@
+cabal-version:      2.2
+name:               timeline-core
+version:            0.1.0.0
+synopsis:
+  Core types and functions for use within other timeline-* packages
+
+license:            BSD-3-Clause
+license-file:       LICENSE
+author:             Bellroy Tech Team <haskell@bellroy.com>
+maintainer:         Bellroy Tech Team <haskell@bellroy.com>
+category:           Development
+build-type:         Simple
+tested-with:        GHC ==8.10.7 || ==9.2.6 || ==9.4.4
+extra-source-files:
+  CHANGELOG.md
+  README.md
+
+source-repository head
+  type:     git
+  location: https://github.com/bellroy/timeline.git
+
+library
+  hs-source-dirs:   src/
+  exposed-modules:  Data.Timeline.Internal
+  default-language: Haskell2010
+  ghc-options:      -Wall -fno-warn-unused-do-bind -fno-warn-unused-imports
+  build-depends:
+    , base                 >=4.14.3  && <4.18
+    , containers           >=0.6.5   && <0.7
+    , indexed-traversable  >=0.1.2   && <0.2
+    , semigroupoids        >=5.3.7   && <5.4
+    , template-haskell     >=2.16.0  && <2.20
+    , text                 >=1.2.4.1 && <2.1
+    , th-compat            >=0.1.4   && <0.2
+    , time                 >=1.9.3   && <1.13
