+++ timeline-hedgehog/timeline-hedgehog.cabal
@@ -19,10 +19,10 @@ source-repository head
 
 common deps
   build-depends:
-    , base      >=4.14.3  && <4.18
-    , hedgehog  >=1.1     && <1.3
-    , time      >=1.9.3   && <1.13
-    , timeline  ==0.1.0.0
+    , base           >=4.14.3  && <4.18
+    , hedgehog       >=1.1     && <1.3
+    , time           >=1.9.3   && <1.13
+    , timeline-core  ==0.1.0.0
 
 library
   import:           deps
