--- timeline-tests/README.md
@@ -1,3 +1,0 @@
-# timeline
-
-Provides a container type `Timeline a` for handling data that changes over time.
