--- timeline-tests/timeline-tests.cabal
@@ -1,44 +1,0 @@
-cabal-version:      2.2
-name:               timeline-tests
-version:            0.1.0.0
-synopsis:           Tests for the timeline library
-license:            BSD-3-Clause
-license-file:       LICENSE
-author:             Bellroy Tech Team <haskell@bellroy.com>
-maintainer:         Bellroy Tech Team <haskell@bellroy.com>
-category:           Development
-build-type:         Simple
-tested-with:        GHC ==8.10.7 || ==9.2.6 || ==9.4.4
-extra-source-files:
-  CHANGELOG.md
-  README.md
-  test/golden/*.txt
-
-source-repository head
-  type:     git
-  location: https://github.com/bellroy/timeline.git
-
-library
-
-test-suite tests
-  type:               exitcode-stdio-1.0
-  hs-source-dirs:     test
-  main-is:            Main.hs
-  other-modules:      Data.TimelineTest
-  build-tool-depends: tasty-discover:tasty-discover
-  build-depends:
-    , base                 >=4.14.3    && <4.18
-    , bytestring           >=0.10      && <0.12
-    , containers           >=0.6.5     && <0.7
-    , hashable             ^>=1.4.2.0
-    , hedgehog             >=1.1       && <1.3
-    , indexed-traversable  ^>=0.1.2
-    , tasty                ^>=1.4.3
-    , tasty-golden         ^>=2.3.5
-    , tasty-hedgehog       >=1.2.0.0
-    , tasty-hunit          ^>=0.10.0.3
-    , text                 >=1.2.4.1   && <2.1
-    , time                 >=1.9.3     && <1.13
-    , timeline             ==0.1.0.0
-    , timeline-hedgehog    ==0.1.0.0
-    , transformers         ^>=0.5.6.2
