+++ timeline/test/golden/all_non-overlapping_situations_together.txt
