+++ timeline/test/golden/two_groups_of_overlap.txt
