+++ timeline/timeline.cabal
@@ -36,3 +36,27 @@ library
   exposed-modules:  Data.Timeline
   default-language: Haskell2010
   ghc-options:      -Wall -fno-warn-unused-do-bind -fno-warn-unused-imports
+  build-depends:    timeline-core ==0.1.0.0
+
+test-suite tests
+  type:               exitcode-stdio-1.0
+  hs-source-dirs:     test
+  main-is:            Main.hs
+  other-modules:      Data.TimelineTest
+  build-tool-depends: tasty-discover:tasty-discover
+  build-depends:
+    , base                 >=4.14.3    && <4.18
+    , bytestring           >=0.10      && <0.12
+    , containers           >=0.6.5     && <0.7
+    , hashable             ^>=1.4.2.0
+    , hedgehog             >=1.1       && <1.3
+    , indexed-traversable  ^>=0.1.2
+    , tasty                ^>=1.4.3
+    , tasty-golden         ^>=2.3.5
+    , tasty-hedgehog       >=1.2.0.0
+    , tasty-hunit          ^>=0.10.0.3
+    , text                 >=1.2.4.1   && <2.1
+    , time                 >=1.9.3     && <1.13
+    , timeline
+    , timeline-hedgehog    ==0.1.0.0
+    , transformers         ^>=0.5.6.2
